+++ test/test_query.py
@@ -1,13 +1,16 @@
-from liara.nodes import DocumentNode, NodeKind, Node
+from liara.nodes import DocumentNode, StaticNode, Node
 import pathlib
 
 
 class MockDocumentNode(DocumentNode):
     def __init__(self, src, metadata):
-        self.kind = NodeKind.Document
+        super().__init__(src, pathlib.PurePosixPath(src))
         self.metadata = metadata
-        self.src = src
-        self.path = pathlib.PurePosixPath(src)
+
+
+class MockStaticNode(StaticNode):
+    def __init__(self, src):
+        super().__init__(src, pathlib.PurePosixPath(src))
 
 
 def test_query_filter_by_tag():
@@ -23,3 +26,22 @@ def test_query_filter_by_tag():
 
     assert len(with_a) == 2
     assert len(with_b) == 1
+
+
+def test_query_filter_by_kind():
+    n1 = MockDocumentNode('/a', {})
+    n2 = MockStaticNode('/b')
+
+    root = Node()
+    root.path = pathlib.PurePosixPath('/')
+    root.add_child(n1)
+    root.add_child(n2)
+
+    doc_children = root.select_children().with_node_kinds('document')
+    assert len(list(doc_children)) == 1
+
+    static_children = root.select_children().with_node_kinds('static')
+    assert len(list(static_children)) == 1
+
+    doc_children = root.select_children().without_node_kinds('static')
+    assert len(list(doc_children)) == 1